Add FutureToObservable
